Test st-daemon-monitor log messages

+++ nlw/t/daemon-monitor.t
@@ -2,7 +2,7 @@
 # @COPYRIGHT@
 use warnings;
 use strict;
-use Test::More tests => 66;
+use Test::More tests => 80;
 use File::Temp qw/tempfile/;
 
 ok -x 'bin/st-daemon-monitor', "it's executable";
@@ -38,10 +38,14 @@ sub reap {
     die "failed to reap $pid";
 }
 
+my $last_log;
+
 sub test_monitor ($$;$) {
     my $cranky = shift;
     my $mon_args = shift;
     my $process_lives = shift || undef;
+
+    local $Test::Builder::Level = $Test::Builder::Level + 1;
 
     unlink 't/tmp/mon';
     pass "begin ($cranky; $mon_args)";
@@ -58,8 +62,12 @@ sub test_monitor ($$;$) {
         sleep 1;
     }
 
-    my $rc = system("bin/st-daemon-monitor ".$mon_args.
-            qq{ --pidfile $pidfile --init "/bin/touch t/tmp/mon"});
+    my ($logfh,$logfile) = tempfile();
+    my $rc = system(
+        "bin/st-daemon-monitor ".$mon_args.
+        qq{ --pidfile $pidfile --init "/bin/touch t/tmp/mon"}.
+        qq{>>$logfile}
+    );
     my $exit = $rc >> 8;
     if ($process_lives) {
         is $exit, 0, "didn't kill the daemon";
@@ -70,14 +78,30 @@ sub test_monitor ($$;$) {
         ok -f 't/tmp/mon', "ran the init cmd";
     }
 
-    for (1..5) {
-        last if waitpid($cranky_pid,1)==$cranky_pid; # non-blocking
-        kill 9, $cranky_pid;
-        last if waitpid($cranky_pid,1)==$cranky_pid; # non-blocking
-        diag "waiting for cranky...";
-        sleep 1;
+    if ($cranky !~ m#^/bin/true#) {
+        for (1..5) {
+            last if waitpid($cranky_pid,1)==$cranky_pid; # non-blocking
+            kill 9, $cranky_pid;
+            last if waitpid($cranky_pid,1)==$cranky_pid; # non-blocking
+            diag "waiting for daemon...";
+            sleep 1;
+        }
     }
     pass "done ($cranky; $mon_args)";
+
+    seek $logfh, 0, 0;
+    $last_log = '';
+    {
+        local $/;
+        ($last_log) = <$logfh>;
+    }
+    close $logfh;
+}
+
+sub logged_like ($) {
+    my $re = shift;
+    local $Test::Builder::Level = $Test::Builder::Level + 1;
+    like $last_log, qr/$re/m, 'logged correctly';
 }
 
 my $c = 'dev-bin/cranky.pl ';
@@ -86,35 +110,47 @@ test_monitor('/bin/sleep 5', ''                           => 'lives');
 test_monitor('/bin/sleep 5', '--rss 32 --vsz 32 --fds 10' => 'lives');
 
 test_monitor('/bin/true',''); # special case: harness will reap
+logged_like qr/is (gone|zombified)/;
 
 test_monitor($c.'--ram 64',         '--rss 64');
-test_monitor($c.'--ram 64',         '--vsz 128');
+logged_like qr/is too big \(RSS\)/i;
+test_monitor($c.'--ram 64',         '--vsz 256' => 'lives');
+test_monitor($c.'--ram 64',         '--vsz 64');
+logged_like qr/is too big \(vsize\)/i;
 test_monitor($c.'--ram 64 --fds 64','--vsz 256 --fds 32');
+logged_like qr/has too many files open/i;
 test_monitor($c.'--fds 64',         '--vsz 256 --fds 32');
+logged_like qr/has too many files open/i;
 
 socket_tests: {
     # check for open port only; monitor doesn't try to connect
+    my $port = $>+26000;
 
-    test_monitor($c.'--after 5 --serv none', ''                  => 'lives');
-    test_monitor($c.'--after 5 --serv none', '--tcp '.($>+26000)           );
-    test_monitor($c.'--after 5',             '--tcp '.($>+26000) => 'lives');
+    test_monitor($c.'--after 5 --serv none', ''            => 'lives');
+    test_monitor($c.'--after 5 --serv none', "--tcp $port"           );
+    logged_like qr/tcp port $port is not open/i;
+    test_monitor($c.'--after 5',             "--tcp $port" => 'lives');
 
     # socket timeout
     test_monitor(
         $c.'--serv stall',
-        '--scgi --tcp '.($>+26000)
+        "--scgi --tcp $port"
     );
+    logged_like qr/timeout/;
 
     # cranky should give a 403
     test_monitor(
         $c.'--serv scgi',
-        '--scgi --tcp '.($>+26000)
+        "--scgi --tcp $port"
     );
+    logged_like qr/non-200/;
 
     # cranky gives a 200
     test_monitor(
         $c.'--after 5 --serv scgi --scgi ok',
-        '--scgi --tcp '.($>+26000)
+        "--scgi --tcp $port"
         => 'lives'
     );
 }
+
+pass "all done";
